Update like.controller.js
Controller for toggling like for tweet comment and for videos is complete in this commit. The user can now like and unlike a video, tweet, or comment.

If a user likes a video, tweet, or comment, a new Like object is created in the database with the user's ID and the ID of the video, tweet, or comment. If a user unlikes a video, tweet, or comment, the existing Like object is deleted from the database.

The user can also get all of the videos they have liked. This is done by finding all of the Like objects in the database with the user's ID and then populating the video field of each Like object.