Update subscription.controller.js
In the subscription controller, I created new functions for toggleSubscription, getUserChannelSubscribers, and getSubscribedChannels. These functions will allow users to toggle their subscription status to channels, get a list of channels they're subscribed to, and return a list of subscribers for a given channel.  The getUserChannelSubscribers and getSubscribedChannels functions return the results in JSON format with a status code of 200 and an appropriate success message. Any errors that occur while fetching the list of subscribers or subscribed channels will be caught and an error message returned.