Update tweet.controller.js
In the tweet controller, I created functions for creating a tweet, getting a user's tweets, updating a tweet, and deleting a tweet. The createTweet function allows a user to create a new tweet, while getUserTweets retrieves all of a user's tweets. The updateTweet function enables a user to edit an existing tweet, and the deleteTweet function deletes a tweet. These functions are essential for managing tweets within the application.